pyproject.toml
[tool.poetry]
name = "ts-preprocessing"
version = "0.1.0"
description = "A preprocessing library for foundation time series models"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "preprocessing"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
torchvision = "^0.17.0"
huggingface-hub = "^0.20.3"
numpy = "^1.26.0"
pandas = "^2.2.0"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
hypothesis = "^6.98.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
download-data = "scripts.download_data:main"
run-preprocessing = "scripts.run_preprocessing:main"
run-augmentation = "scripts.run_augmentation:main" 